@@ -4,22 +4,28 @@ declare(strict_types=1);
 
 namespace Yiisoft\Validator\Rule;
 
-use Yiisoft\Validator\HasValidationErrorMessage;
+use Yiisoft\Validator\FormatterInterface;
 use Yiisoft\Validator\Result;
 use Yiisoft\Validator\Rule;
 use Yiisoft\Validator\RuleSet;
 use Yiisoft\Validator\ValidationContext;
 
 /**
- * GroupRule validates a single value for a set of custom rules
+ * Validates a single value for a set of custom rules.
  */
 abstract class GroupRule extends Rule
 {
-    use HasValidationErrorMessage;
+    public function __construct(
+        protected string $message = 'This value is not a valid.',
+        ?FormatterInterface $formatter = null,
+        bool $skipOnEmpty = false,
+        bool $skipOnError = false,
+        $when = null
+    ) {
+        parent::__construct(formatter: $formatter, skipOnEmpty: $skipOnEmpty, skipOnError: $skipOnError, when: $when);
+    }
 
-    protected string $message = 'This value is not a valid.';
-
-    protected function validateValue($value, ValidationContext $context = null): Result
+    protected function validateValue($value, ?ValidationContext $context = null): Result
     {
         $result = new Result();
         if (!$this->getRuleSet()->validate($value, $context)->isValid()) {
@@ -31,8 +37,6 @@ abstract class GroupRule extends Rule
 
     /**
      * Return custom rules set
-     *
-     * @return RuleSet
      */
     abstract protected function getRuleSet(): RuleSet;
 
