@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace Yiisoft\Validator\Rule;
 
-use Yiisoft\Validator\HasValidationErrorMessage;
+use Attribute;
+use RuntimeException;
+use Yiisoft\Validator\FormatterInterface;
 use Yiisoft\Validator\Result;
 use Yiisoft\Validator\Rule;
 use Yiisoft\Validator\ValidationContext;
@@ -13,45 +15,50 @@ use function is_string;
 use function strlen;
 
 /**
- * UrlValidator validates that the attribute value is a valid HTTP or HTTPS URL.
+ * Validates that the value is a valid HTTP or HTTPS URL.
  *
- * Note that this validator only checks if the URL scheme and host part are correct.
+ * Note that this rule only checks if the URL scheme and host part are correct.
  * It does not check the remaining parts of a URL.
  */
+#[Attribute(Attribute::TARGET_PROPERTY)]
 final class Url extends Rule
 {
-    use HasValidationErrorMessage;
+    public function __construct(
+        /**
+         * @var string the regular expression used to validate the value.
+         * The pattern may contain a `{schemes}` token that will be replaced
+         * by a regular expression which represents the {@see $schemes}.
+         *
+         * Note that if you want to reuse the pattern in HTML5 input it should have ^ and $, should not have any
+         * modifiers and should not be case-insensitive.
+         */
+        private string $pattern = '/^{schemes}:\/\/(([a-zA-Z0-9][a-zA-Z0-9_-]*)(\.[a-zA-Z0-9][a-zA-Z0-9_-]*)+)(?::\d{1,5})?([?\/#].*$|$)/',
+        /**
+         * @var array list of URI schemes which should be considered valid. By default, http and https
+         * are considered to be valid schemes.
+         */
+        private array $validSchemes = ['http', 'https'],
+        /**
+         * @var bool whether validation process should take into account IDN (internationalized
+         * domain names). Defaults to false meaning that validation of URLs containing IDN will always
+         * fail. Note that in order to use IDN validation you have to install and enable `intl` PHP
+         * extension, otherwise an exception would be thrown.
+         */
+        private bool $enableIDN = false,
+        private string $message = 'This value is not a valid URL.',
+        ?FormatterInterface $formatter = null,
+        bool $skipOnEmpty = false,
+        bool $skipOnError = false,
+        $when = null
+    ) {
+        parent::__construct(formatter: $formatter, skipOnEmpty: $skipOnEmpty, skipOnError: $skipOnError, when: $when);
 
-    /**
-     * @var string the regular expression used to validateValue the attribute value.
-     * The pattern may contain a `{schemes}` token that will be replaced
-     * by a regular expression which represents the {@see schemes()}.
-     *
-     * Note that if you want to reuse the pattern in HTML5 input it should have ^ and $, should not have any modifiers
-     * and should not be case-insensitive.
-     */
-    private string $pattern = '/^{schemes}:\/\/(([a-zA-Z0-9][a-zA-Z0-9_-]*)(\.[a-zA-Z0-9][a-zA-Z0-9_-]*)+)(?::\d{1,5})?([?\/#].*$|$)/';
-    /**
-     * @var array list of URI schemes which should be considered valid. By default, http and https
-     * are considered to be valid schemes.
-     */
-    private array $validSchemes = ['http', 'https'];
-    /**
-     * @var bool whether validation process should take into account IDN (internationalized
-     * domain names). Defaults to false meaning that validation of URLs containing IDN will always
-     * fail. Note that in order to use IDN validation you have to install and enable `intl` PHP
-     * extension, otherwise an exception would be thrown.
-     */
-    private bool $enableIDN = false;
-
-    private string $message = 'This value is not a valid URL.';
-
-    public static function rule(): self
-    {
-        return new self();
+        if ($enableIDN && !function_exists('idn_to_ascii')) {
+            throw new RuntimeException('In order to use IDN validation intl extension must be installed and enabled.');
+        }
     }
 
-    protected function validateValue($value, ValidationContext $context = null): Result
+    protected function validateValue($value, ?ValidationContext $context = null): Result
     {
         $result = new Result();
 
@@ -100,41 +107,13 @@ final class Url extends Rule
         return $this->pattern;
     }
 
-    public function pattern(string $pattern): self
-    {
-        $new = clone $this;
-        $new->pattern = $pattern;
-        return $new;
-    }
-
-    public function enableIDN(): self
-    {
-        if (!function_exists('idn_to_ascii')) {
-            throw new \RuntimeException('In order to use IDN validation intl extension must be installed and enabled.');
-        }
-
-        $new = clone $this;
-        $new->enableIDN = true;
-        return $new;
-    }
-
-    public function schemes(array $schemes): self
-    {
-        $new = clone $this;
-        $new->validSchemes = $schemes;
-        return $new;
-    }
-
     public function getOptions(): array
     {
-        return array_merge(
-            parent::getOptions(),
-            [
-                'message' => $this->formatMessage($this->message),
-                'enableIDN' => $this->enableIDN,
-                'validSchemes' => $this->validSchemes,
-                'pattern' => $this->pattern,
-            ],
-        );
+        return array_merge(parent::getOptions(), [
+            'pattern' => $this->pattern,
+            'validSchemes' => $this->validSchemes,
+            'enableIDN' => $this->enableIDN,
+            'message' => $this->formatMessage($this->message),
+        ]);
     }
 }
