@@ -7,6 +7,7 @@ namespace Yiisoft\Validator;
 use InvalidArgumentException;
 use Yiisoft\Validator\Rule\Callback;
 use function get_class;
+use function gettype;
 use function is_callable;
 
 /**
@@ -15,8 +16,6 @@ use function is_callable;
 final class RuleSet
 {
     public const PARAMETER_PREVIOUS_RULES_ERRORED = 'previousRulesErrored';
-
-    private ?FormatterInterface $formatter = null;
 
     /**
      * @var RuleInterface[]
@@ -35,18 +34,14 @@ final class RuleSet
      */
     public function add($rule): void
     {
-        $rule = $this->normalizeRule($rule);
-        if ($this->formatter !== null && $rule instanceof FormattableRuleInterface) {
-            $rule = $rule->withFormatter($this->formatter);
-        }
-        $this->rules[] = $rule;
+        $this->rules[] = $this->normalizeRule($rule);
     }
 
     public function validate($value, ValidationContext $context = null): Result
     {
         $context = $context ?? new ValidationContext();
-        $compoundResult = new Result();
 
+        $compoundResult = new Result();
         foreach ($this->rules as $rule) {
             $ruleResult = $rule->validate($value, $context);
             if ($ruleResult->isValid()) {
@@ -59,14 +54,13 @@ final class RuleSet
                 $compoundResult->addError($error->getMessage(), $error->getValuePath());
             }
         }
-
         return $compoundResult;
     }
 
     private function normalizeRule($rule): RuleInterface
     {
         if (is_callable($rule)) {
-            $rule = Callback::rule($rule);
+            return new Callback($rule);
         }
 
         if (!$rule instanceof RuleInterface) {
@@ -78,14 +72,6 @@ final class RuleSet
         }
 
         return $rule;
-    }
-
-    public function withFormatter(?FormatterInterface $formatter): self
-    {
-        $new = clone $this;
-        $new->formatter = $formatter;
-        $new->addFormatterToRules($formatter);
-        return $new;
     }
 
     /**
@@ -104,14 +90,5 @@ final class RuleSet
             }
         }
         return $arrayOfRules;
-    }
-
-    private function addFormatterToRules(?FormatterInterface $formatter): void
-    {
-        foreach ($this->rules as &$rule) {
-            if ($rule instanceof FormattableRuleInterface) {
-                $rule = $rule->withFormatter($formatter);
-            }
-        }
     }
 }
