@@ -14,13 +14,6 @@ use function is_object;
  */
 final class Validator implements ValidatorInterface
 {
-    private ?FormatterInterface $formatter;
-
-    public function __construct(?FormatterInterface $formatter = null)
-    {
-        $this->formatter = $formatter;
-    }
-
     /**
      * @param DataSetInterface|mixed|RulesProviderInterface $data
      * @param Rule[][] $rules
@@ -38,14 +31,8 @@ final class Validator implements ValidatorInterface
 
         foreach ($rules as $attribute => $attributeRules) {
             $ruleSet = new RuleSet($attributeRules);
-            if ($this->formatter !== null) {
-                $ruleSet = $ruleSet->withFormatter($this->formatter);
-            }
+            $tempResult = $ruleSet->validate($data->getAttributeValue($attribute), $context->withAttribute($attribute));
 
-            $tempResult = $ruleSet->validate(
-                $data->getAttributeValue($attribute),
-                $context->withAttribute($attribute)
-            );
             foreach ($tempResult->getErrors() as $error) {
                 $result->addError($error->getMessage(), [$attribute, ...$error->getValuePath()]);
             }
@@ -58,13 +45,6 @@ final class Validator implements ValidatorInterface
         return $result;
     }
 
-    public function withFormatter(?FormatterInterface $formatter): self
-    {
-        $new = clone $this;
-        $new->formatter = $formatter;
-        return $new;
-    }
-
     private function normalizeDataSet($data): DataSetInterface
     {
         if ($data instanceof DataSetInterface) {
@@ -72,7 +52,7 @@ final class Validator implements ValidatorInterface
         }
 
         if (is_object($data) || is_array($data)) {
-            return new ArrayDataSet((array)$data);
+            return new ArrayDataSet((array) $data);
         }
 
         return new ScalarDataSet($data);
