@@ -15,57 +15,49 @@ class HasLengthTest extends TestCase
 {
     public function testValidate(): void
     {
-        $val = HasLength::rule();
-        $this->assertFalse($val->validate(['not a string'])->isValid());
-        $this->assertFalse($val->validate(new \stdClass())->isValid());
-        $this->assertTrue($val->validate('Just some string')->isValid());
-        $this->assertFalse($val->validate(true)->isValid());
-        $this->assertFalse($val->validate(false)->isValid());
+        $rule = new HasLength();
+        $this->assertFalse($rule->validate(['not a string'])->isValid());
+        $this->assertFalse($rule->validate(new \stdClass())->isValid());
+        $this->assertTrue($rule->validate('Just some string')->isValid());
+        $this->assertFalse($rule->validate(true)->isValid());
+        $this->assertFalse($rule->validate(false)->isValid());
     }
 
     public function testValidateLength(): void
     {
-        $val = HasLength::rule()
-            ->min(25)
-            ->max(25);
-        $this->assertTrue($val->validate(str_repeat('x', 25))->isValid());
-        $this->assertTrue($val->validate(str_repeat('€', 25))->isValid());
-        $this->assertFalse($val->validate(str_repeat('x', 125))->isValid());
-        $this->assertFalse($val->validate('')->isValid());
+        $rule = new HasLength(min: 25, max: 25);
+        $this->assertTrue($rule->validate(str_repeat('x', 25))->isValid());
+        $this->assertTrue($rule->validate(str_repeat('€', 25))->isValid());
+        $this->assertFalse($rule->validate(str_repeat('x', 125))->isValid());
+        $this->assertFalse($rule->validate('')->isValid());
 
-        $val = HasLength::rule()
-            ->min(25);
-        $this->assertTrue($val->validate(str_repeat('x', 125))->isValid());
-        $this->assertTrue($val->validate(str_repeat('€', 25))->isValid());
-        $this->assertFalse($val->validate(str_repeat('x', 13))->isValid());
-        $this->assertFalse($val->validate('')->isValid());
+        $rule = new HasLength(min: 25);
+        $this->assertTrue($rule->validate(str_repeat('x', 125))->isValid());
+        $this->assertTrue($rule->validate(str_repeat('€', 25))->isValid());
+        $this->assertFalse($rule->validate(str_repeat('x', 13))->isValid());
+        $this->assertFalse($rule->validate('')->isValid());
 
-        $val = HasLength::rule()
-            ->max(25);
-        $this->assertTrue($val->validate(str_repeat('x', 25))->isValid());
-        $this->assertTrue($val->validate(str_repeat('Ä', 24))->isValid());
-        $this->assertfalse($val->validate(str_repeat('x', 1250))->isValid());
-        $this->assertTrue($val->validate('')->isValid());
+        $rule = new HasLength(max: 25);
+        $this->assertTrue($rule->validate(str_repeat('x', 25))->isValid());
+        $this->assertTrue($rule->validate(str_repeat('Ä', 24))->isValid());
+        $this->assertfalse($rule->validate(str_repeat('x', 1250))->isValid());
+        $this->assertTrue($rule->validate('')->isValid());
 
-        $val = HasLength::rule()
-            ->min(10)
-            ->max(25);
-        $this->assertTrue($val->validate(str_repeat('x', 15))->isValid());
-        $this->assertTrue($val->validate(str_repeat('x', 10))->isValid());
-        $this->assertTrue($val->validate(str_repeat('x', 20))->isValid());
-        $this->assertTrue($val->validate(str_repeat('x', 25))->isValid());
-        $this->assertFalse($val->validate(str_repeat('x', 5))->isValid());
-        $this->assertFalse($val->validate('')->isValid());
+        $rule = new HasLength(min: 10, max: 25);
+        $this->assertTrue($rule->validate(str_repeat('x', 15))->isValid());
+        $this->assertTrue($rule->validate(str_repeat('x', 10))->isValid());
+        $this->assertTrue($rule->validate(str_repeat('x', 20))->isValid());
+        $this->assertTrue($rule->validate(str_repeat('x', 25))->isValid());
+        $this->assertFalse($rule->validate(str_repeat('x', 5))->isValid());
+        $this->assertFalse($rule->validate('')->isValid());
     }
 
     public function testValidateMin(): void
     {
-        $rule = HasLength::rule()
-            ->min(1);
-
+        $rule = new HasLength(min: 1);
         $result = $rule->validate('');
-        $this->assertFalse($result->isValid());
 
+        $this->assertFalse($result->isValid());
         $this->assertEquals(
             ['This value should contain at least {min, number} {min, plural, one{character} other{characters}}.'],
             $result->getErrorMessages()
@@ -75,9 +67,7 @@ class HasLengthTest extends TestCase
 
     public function testValidateMax(): void
     {
-        $rule = HasLength::rule()
-            ->max(100);
-
+        $rule = new HasLength(max: 100);
         $this->assertTrue($rule->validate(str_repeat('x', 5))->isValid());
 
         $result = $rule->validate(str_repeat('x', 1230));
@@ -90,33 +80,34 @@ class HasLengthTest extends TestCase
 
     public function testValidateMessages()
     {
-        $rule = HasLength::rule()
-            ->message('is not string error')
-            ->tooShortMessage('is to short test')
-            ->tooLongMessage('is to long test')
-            ->min(3)
-            ->max(5);
+        $rule = new HasLength(
+            min: 3,
+            max: 5,
+            message: 'is not string error',
+            tooShortMessage: 'is too short test',
+            tooLongMessage: 'is too long test'
+        );
 
         $this->assertEquals(['is not string error'], $rule->validate(null)->getErrorMessages());
-        $this->assertEquals(['is to short test'], $rule->validate(str_repeat('x', 1))->getErrorMessages());
-        $this->assertEquals(['is to long test'], $rule->validate(str_repeat('x', 6))->getErrorMessages());
+        $this->assertEquals(['is too short test'], $rule->validate(str_repeat('x', 1))->getErrorMessages());
+        $this->assertEquals(['is too long test'], $rule->validate(str_repeat('x', 6))->getErrorMessages());
     }
 
     public function testName(): void
     {
-        $this->assertEquals('hasLength', HasLength::rule()->getName());
+        $this->assertEquals('hasLength', (new HasLength())->getName());
     }
 
     public function optionsProvider(): array
     {
         return [
             [
-                HasLength::rule(),
+                new HasLength(),
                 [
-                    'message' => 'This value must be a string.',
                     'min' => null,
-                    'tooShortMessage' => 'This value should contain at least {min, number} {min, plural, one{character} other{characters}}.',
                     'max' => null,
+                    'message' => 'This value must be a string.',
+                    'tooShortMessage' => 'This value should contain at least {min, number} {min, plural, one{character} other{characters}}.',
                     'tooLongMessage' => 'This value should contain at most {max, number} {max, plural, one{character} other{characters}}.',
                     'encoding' => 'UTF-8',
                     'skipOnEmpty' => false,
@@ -124,12 +115,12 @@ class HasLengthTest extends TestCase
                 ],
             ],
             [
-                HasLength::rule()->min(3),
+                new HasLength(min: 3),
                 [
-                    'message' => 'This value must be a string.',
                     'min' => 3,
-                    'tooShortMessage' => 'This value should contain at least {min, number} {min, plural, one{character} other{characters}}.',
                     'max' => null,
+                    'message' => 'This value must be a string.',
+                    'tooShortMessage' => 'This value should contain at least {min, number} {min, plural, one{character} other{characters}}.',
                     'tooLongMessage' => 'This value should contain at most {max, number} {max, plural, one{character} other{characters}}.',
                     'encoding' => 'UTF-8',
                     'skipOnEmpty' => false,
@@ -137,12 +128,12 @@ class HasLengthTest extends TestCase
                 ],
             ],
             [
-                HasLength::rule()->max(3),
+                new HasLength(max: 3),
                 [
-                    'message' => 'This value must be a string.',
                     'min' => null,
-                    'tooShortMessage' => 'This value should contain at least {min, number} {min, plural, one{character} other{characters}}.',
                     'max' => 3,
+                    'message' => 'This value must be a string.',
+                    'tooShortMessage' => 'This value should contain at least {min, number} {min, plural, one{character} other{characters}}.',
                     'tooLongMessage' => 'This value should contain at most {max, number} {max, plural, one{character} other{characters}}.',
                     'encoding' => 'UTF-8',
                     'skipOnEmpty' => false,
@@ -150,12 +141,12 @@ class HasLengthTest extends TestCase
                 ],
             ],
             [
-                HasLength::rule()->min(3)->max(4)->encoding('windows-1251'),
+                new HasLength(min: 3, max: 4, encoding: 'windows-1251'),
                 [
-                    'message' => 'This value must be a string.',
                     'min' => 3,
-                    'tooShortMessage' => 'This value should contain at least {min, number} {min, plural, one{character} other{characters}}.',
                     'max' => 4,
+                    'message' => 'This value must be a string.',
+                    'tooShortMessage' => 'This value should contain at least {min, number} {min, plural, one{character} other{characters}}.',
                     'tooLongMessage' => 'This value should contain at most {max, number} {max, plural, one{character} other{characters}}.',
                     'encoding' => 'windows-1251',
                     'skipOnEmpty' => false,
