@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Validator\Tests\Rule;
 
+use ArrayObject;
 use PHPUnit\Framework\TestCase;
 use Yiisoft\Validator\Rule;
 use Yiisoft\Validator\Rule\InRange;
@@ -15,20 +16,22 @@ class InRangeTest extends TestCase
 {
     public function testValidate(): void
     {
-        $val = InRange::rule(range(1, 10));
-        $this->assertTrue($val->validate(1)->isValid());
-        $this->assertFalse($val->validate(0)->isValid());
-        $this->assertFalse($val->validate(11)->isValid());
-        $this->assertFalse($val->validate(5.5)->isValid());
-        $this->assertTrue($val->validate(10)->isValid());
-        $this->assertTrue($val->validate('10')->isValid());
-        $this->assertTrue($val->validate('5')->isValid());
+        $rule = new InRange(range(1, 10));
+
+        $this->assertTrue($rule->validate(1)->isValid());
+        $this->assertFalse($rule->validate(0)->isValid());
+        $this->assertFalse($rule->validate(11)->isValid());
+        $this->assertFalse($rule->validate(5.5)->isValid());
+        $this->assertTrue($rule->validate(10)->isValid());
+        $this->assertTrue($rule->validate('10')->isValid());
+        $this->assertTrue($rule->validate('5')->isValid());
     }
 
     public function testValidateEmpty(): void
     {
-        $rule = InRange::rule(range(10, 20))->skipOnEmpty(false);
-        $this->assertFalse($rule->validate(null)->isValid()); //row RangeValidatorTest.php:101
+        $rule = new InRange(range(10, 20));
+
+        $this->assertFalse($rule->validate(null)->isValid()); // row RangeValidatorTest.php:101
         $this->assertFalse($rule->validate('0')->isValid());
         $this->assertFalse($rule->validate(0)->isValid());
         $this->assertFalse($rule->validate('')->isValid());
@@ -37,21 +40,17 @@ class InRangeTest extends TestCase
     public function testValidateArrayValue(): void
     {
         // Test in array, values are arrays. IE: ['a'] in [['a'], ['b']]
-        $rule = InRange::rule([['a'], ['b']]);
-
+        $rule = new InRange([['a'], ['b']]);
         $this->assertTrue($rule->validate(['a'])->isValid());
 
-
-        // Test range as ArrayObject.
-        $rule = InRange::rule(new \ArrayObject(['a', 'b']));
-
+        // Test range as ArrayObject
+        $rule = new InRange(new ArrayObject(['a', 'b']));
         $this->assertTrue($rule->validate('a')->isValid());
     }
 
     public function testValidateStrict(): void
     {
-        $rule = InRange::rule(range(1, 10))
-            ->strict();
+        $rule = new InRange(range(1, 10), strict: true);
 
         $this->assertTrue($rule->validate(1)->isValid());
         $this->assertTrue($rule->validate(5)->isValid());
@@ -63,8 +62,7 @@ class InRangeTest extends TestCase
 
     public function testValidateArrayValueStrict(): void
     {
-        $rule = InRange::rule(range(1, 10, 1))
-            ->strict();
+        $rule = new InRange(range(1, 10), strict: true);
 
         $this->assertFalse($rule->validate(['1', '2', '3', '4', '5', '6'])->isValid());
         $this->assertFalse($rule->validate(['1', '2', '3', 4, 5, 6])->isValid());
@@ -72,8 +70,7 @@ class InRangeTest extends TestCase
 
     public function testValidateNot()
     {
-        $rule = InRange::rule(range(1, 10, 1))
-            ->not();
+        $rule = new InRange(range(1, 10), not: true);
 
         $this->assertFalse($rule->validate(1)->isValid());
         $this->assertTrue($rule->validate(0)->isValid());
@@ -86,41 +83,42 @@ class InRangeTest extends TestCase
 
     public function testName(): void
     {
-        $this->assertEquals('inRange', InRange::rule(range(1, 10))->getName());
+        $rule = new InRange(range(1, 10));
+        $this->assertEquals('inRange', $rule->getName());
     }
 
     public function optionsProvider(): array
     {
         return [
             [
-                InRange::rule(range(1, 10)),
+                new InRange(range(1, 10)),
                 [
-                    'message' => 'This value is invalid.',
                     'range' => range(1, 10),
                     'strict' => false,
                     'not' => false,
+                    'message' => 'This value is invalid.',
                     'skipOnEmpty' => false,
                     'skipOnError' => false,
                 ],
             ],
             [
-                InRange::rule(range(1, 2))->strict(),
+                new InRange(range(1, 2), strict: true),
                 [
-                    'message' => 'This value is invalid.',
                     'range' => [1, 2],
                     'strict' => true,
                     'not' => false,
+                    'message' => 'This value is invalid.',
                     'skipOnEmpty' => false,
                     'skipOnError' => false,
                 ],
             ],
             [
-                InRange::rule(range(1, 2))->not(),
+                new InRange(range(1, 2), not: true),
                 [
-                    'message' => 'This value is invalid.',
                     'range' => [1, 2],
                     'strict' => false,
                     'not' => true,
+                    'message' => 'This value is invalid.',
                     'skipOnEmpty' => false,
                     'skipOnError' => false,
                 ],
