@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Validator\Tests\Rule;
 
 use PHPUnit\Framework\TestCase;
+use RuntimeException;
 use Yiisoft\Validator\Rule;
 use Yiisoft\Validator\Rule\Ip;
 
@@ -17,7 +18,8 @@ class IpTest extends TestCase
     {
         $this->expectException(\RuntimeException::class);
         $this->expectExceptionMessage('Both IPv4 and IPv6 checks can not be disabled at the same time');
-        Ip::rule()->disallowIpv4()->disallowIpv6()->validate('');
+
+        (new Ip(allowIpv4: false, allowIpv6: false))->validate('');
     }
 
     public function provideRangesForSubstitution(): array
@@ -54,21 +56,21 @@ class IpTest extends TestCase
     /**
      * @dataProvider provideRangesForSubstitution
      */
-    public function testRangesSubstitution(array $range, array $expectedRange): void
+    public function testRangesSubstitution(array $ranges, array $expectedRanges): void
     {
-        $validator = Ip::rule()->ranges($range);
-        $this->assertEquals($expectedRange, $validator->getRanges());
+        $rule = new Ip(ranges: $ranges);
+        $this->assertEquals($expectedRanges, $rule->getRanges());
     }
 
     public function testValidateOrder(): void
     {
-        $validator = Ip::rule()->ranges(['10.0.0.1', '!10.0.0.0/8', '!babe::/8', 'any']);
+        $rule = new Ip(ranges: ['10.0.0.1', '!10.0.0.0/8', '!babe::/8', 'any']);
 
-        $this->assertTrue($validator->validate('10.0.0.1')->isValid());
-        $this->assertFalse($validator->validate('10.0.0.2')->isValid());
-        $this->assertTrue($validator->validate('192.168.5.101')->isValid());
-        $this->assertTrue($validator->validate('cafe::babe')->isValid());
-        $this->assertFalse($validator->validate('babe::cafe')->isValid());
+        $this->assertTrue($rule->validate('10.0.0.1')->isValid());
+        $this->assertFalse($rule->validate('10.0.0.2')->isValid());
+        $this->assertTrue($rule->validate('192.168.5.101')->isValid());
+        $this->assertTrue($rule->validate('cafe::babe')->isValid());
+        $this->assertFalse($rule->validate('babe::cafe')->isValid());
     }
 
     public function provideBadIps(): array
@@ -88,416 +90,349 @@ class IpTest extends TestCase
      */
     public function testValidateNotAnIP($badIp): void
     {
-        $validator = Ip::rule();
-        $this->assertFalse($validator->validate($badIp)->isValid());
+        $this->assertFalse((new Ip())->validate($badIp)->isValid());
     }
 
     public function testValidateIPv4(): void
     {
-        $validator = Ip::rule();
+        $rule = new Ip();
+        $this->assertTrue($rule->validate('192.168.10.11')->isValid());
+        $this->assertFalse($rule->validate('192.168.005.001')->isValid()); // leading zero not supported
+        $this->assertFalse($rule->validate('192.168.5.321')->isValid());
+        $this->assertFalse($rule->validate('!192.168.5.32')->isValid());
+        $this->assertFalse($rule->validate('192.168.5.32/11')->isValid());
 
-        $this->assertTrue($validator->validate('192.168.10.11')->isValid());
-        $this->assertFalse($validator->validate('192.168.005.001')->isValid()); // leading zero not supported
-        $this->assertFalse($validator->validate('192.168.5.321')->isValid());
-        $this->assertFalse($validator->validate('!192.168.5.32')->isValid());
-        $this->assertFalse($validator->validate('192.168.5.32/11')->isValid());
+        $rule = new Ip(allowIpv4: false);
+        $this->assertFalse($rule->validate('192.168.10.11')->isValid());
 
-        $validator = $validator->disallowIpv4();
-        $this->assertFalse($validator->validate('192.168.10.11')->isValid());
+        $rule = new Ip(allowSubnet: true);
+        $this->assertTrue($rule->validate('192.168.5.32/11')->isValid());
+        $this->assertTrue($rule->validate('192.168.5.32/32')->isValid());
+        $this->assertTrue($rule->validate('0.0.0.0/0')->isValid());
+        $this->assertFalse($rule->validate('192.168.5.32/33')->isValid());
+        $this->assertFalse($rule->validate('192.168.5.32/33')->isValid());
+        $this->assertFalse($rule->validate('192.168.5.32/af')->isValid());
+        $this->assertFalse($rule->validate('192.168.5.32/11/12')->isValid());
 
-        $validator = $validator->allowIpv4()->allowSubnet();
-        $this->assertTrue($validator->validate('192.168.5.32/11')->isValid());
-        $this->assertTrue($validator->validate('192.168.5.32/32')->isValid());
-        $this->assertTrue($validator->validate('0.0.0.0/0')->isValid());
-        $this->assertFalse($validator->validate('192.168.5.32/33')->isValid());
-        $this->assertFalse($validator->validate('192.168.5.32/33')->isValid());
-        $this->assertFalse($validator->validate('192.168.5.32/af')->isValid());
-        $this->assertFalse($validator->validate('192.168.5.32/11/12')->isValid());
+        $rule = new Ip(requireSubnet: true);
+        $this->assertTrue($rule->validate('10.0.0.1/24')->isValid());
+        $this->assertTrue($rule->validate('10.0.0.1/0')->isValid());
+        $this->assertFalse($rule->validate('10.0.0.1')->isValid());
 
-        $validator = $validator->requireSubnet();
-        $this->assertTrue($validator->validate('10.0.0.1/24')->isValid());
-        $this->assertTrue($validator->validate('10.0.0.1/0')->isValid());
-        $this->assertFalse($validator->validate('10.0.0.1')->isValid());
-
-        $validator = $validator->allowNegation();
-        $this->assertTrue($validator->validate('!192.168.5.32/32')->isValid());
-        $this->assertFalse($validator->validate('!!192.168.5.32/32')->isValid());
+        $rule = new Ip(requireSubnet: true, allowNegation: true);
+        $this->assertTrue($rule->validate('!192.168.5.32/32')->isValid());
+        $this->assertFalse($rule->validate('!!192.168.5.32/32')->isValid());
     }
 
     public function testValidateIPv6(): void
     {
-        $validator = Ip::rule();
+        $rule = new Ip();
+        $this->assertTrue($rule->validate('2008:fa::1')->isValid());
+        $this->assertTrue($rule->validate('2008:00fa::0001')->isValid());
+        $this->assertFalse($rule->validate('2008:fz::0')->isValid());
+        $this->assertFalse($rule->validate('2008:fa::0::1')->isValid());
+        $this->assertFalse($rule->validate('!2008:fa::0::1')->isValid());
+        $this->assertFalse($rule->validate('2008:fa::0:1/64')->isValid());
 
-        $this->assertTrue($validator->validate('2008:fa::1')->isValid());
-        $this->assertTrue($validator->validate('2008:00fa::0001')->isValid());
-        $this->assertFalse($validator->validate('2008:fz::0')->isValid());
-        $this->assertFalse($validator->validate('2008:fa::0::1')->isValid());
-        $this->assertFalse($validator->validate('!2008:fa::0::1')->isValid());
-        $this->assertFalse($validator->validate('2008:fa::0:1/64')->isValid());
+        $rule = new Ip(allowIpv4: false);
+        $this->assertTrue($rule->validate('2008:fa::1')->isValid());
 
-        $validator = $validator->disallowIpv4();
-        $this->assertTrue($validator->validate('2008:fa::1')->isValid());
+        $rule = new Ip(allowIpv4: false, allowSubnet: true);
+        $this->assertTrue($rule->validate('2008:fa::0:1/64')->isValid());
+        $this->assertTrue($rule->validate('2008:fa::0:1/128')->isValid());
+        $this->assertTrue($rule->validate('2008:fa::0:1/0')->isValid());
+        $this->assertFalse($rule->validate('!2008:fa::0:1/0')->isValid());
+        $this->assertFalse($rule->validate('2008:fz::0/129')->isValid());
 
-        $validator = $validator->allowIpv6()->allowSubnet();
-        $this->assertTrue($validator->validate('2008:fa::0:1/64')->isValid());
-        $this->assertTrue($validator->validate('2008:fa::0:1/128')->isValid());
-        $this->assertTrue($validator->validate('2008:fa::0:1/0')->isValid());
-        $this->assertFalse($validator->validate('!2008:fa::0:1/0')->isValid());
-        $this->assertFalse($validator->validate('2008:fz::0/129')->isValid());
+        $rule = new Ip(allowIpv4: false, requireSubnet: true);
+        $this->assertTrue($rule->validate('2008:db0::1/64')->isValid());
+        $this->assertFalse($rule->validate('2008:db0::1')->isValid());
 
-        $validator = $validator->requireSubnet();
-        $this->assertTrue($validator->validate('2008:db0::1/64')->isValid());
-        $this->assertFalse($validator->validate('2008:db0::1')->isValid());
-
-        $validator = $validator->allowNegation();
-        $this->assertTrue($validator->validate('!2008:fa::0:1/64')->isValid());
-        $this->assertFalse($validator->validate('!!2008:fa::0:1/64')->isValid());
+        $rule = new Ip(allowIpv4: false, requireSubnet: true, allowNegation: true);
+        $this->assertTrue($rule->validate('!2008:fa::0:1/64')->isValid());
+        $this->assertFalse($rule->validate('!!2008:fa::0:1/64')->isValid());
     }
 
     public function testValidateIPvBoth(): void
     {
-        $validator = Ip::rule();
+        $rule = new Ip();
+        $this->assertTrue($rule->validate('192.168.10.11')->isValid());
+        $this->assertTrue($rule->validate('2008:fa::1')->isValid());
+        $this->assertTrue($rule->validate('2008:00fa::0001')->isValid());
+        $this->assertFalse($rule->validate('192.168.005.001')->isValid()); // leading zero not allowed
+        $this->assertFalse($rule->validate('192.168.5.321')->isValid());
+        $this->assertFalse($rule->validate('!192.168.5.32')->isValid());
+        $this->assertFalse($rule->validate('192.168.5.32/11')->isValid());
+        $this->assertFalse($rule->validate('2008:fz::0')->isValid());
+        $this->assertFalse($rule->validate('2008:fa::0::1')->isValid());
+        $this->assertFalse($rule->validate('!2008:fa::0::1')->isValid());
+        $this->assertFalse($rule->validate('2008:fa::0:1/64')->isValid());
 
-        $this->assertTrue($validator->validate('192.168.10.11')->isValid());
-        $this->assertTrue($validator->validate('2008:fa::1')->isValid());
-        $this->assertTrue($validator->validate('2008:00fa::0001')->isValid());
-        $this->assertFalse($validator->validate('192.168.005.001')->isValid()); // leading zero not allowed
-        $this->assertFalse($validator->validate('192.168.5.321')->isValid());
-        $this->assertFalse($validator->validate('!192.168.5.32')->isValid());
-        $this->assertFalse($validator->validate('192.168.5.32/11')->isValid());
-        $this->assertFalse($validator->validate('2008:fz::0')->isValid());
-        $this->assertFalse($validator->validate('2008:fa::0::1')->isValid());
-        $this->assertFalse($validator->validate('!2008:fa::0::1')->isValid());
-        $this->assertFalse($validator->validate('2008:fa::0:1/64')->isValid());
+        $rule = new Ip(allowIpv4: false);
+        $this->assertFalse($rule->validate('192.168.10.11')->isValid());
+        $this->assertTrue($rule->validate('2008:fa::1')->isValid());
 
-        $validator = $validator->disallowIpv4();
-        $this->assertFalse($validator->validate('192.168.10.11')->isValid());
-        $this->assertTrue($validator->validate('2008:fa::1')->isValid());
+        $rule = new Ip(allowIpv6: false);
+        $this->assertTrue($rule->validate('192.168.10.11')->isValid());
+        $this->assertFalse($rule->validate('2008:fa::1')->isValid());
 
-        $validator = $validator->disallowIpv6()->allowIpv4();
-        $this->assertTrue($validator->validate('192.168.10.11')->isValid());
-        $this->assertFalse($validator->validate('2008:fa::1')->isValid());
+        $rule = new Ip(requireSubnet: true);
+        $this->assertTrue($rule->validate('192.168.5.32/11')->isValid());
+        $this->assertTrue($rule->validate('192.168.5.32/32')->isValid());
+        $this->assertTrue($rule->validate('0.0.0.0/0')->isValid());
+        $this->assertTrue($rule->validate('2008:fa::0:1/64')->isValid());
+        $this->assertTrue($rule->validate('2008:fa::0:1/128')->isValid());
+        $this->assertTrue($rule->validate('2008:fa::0:1/0')->isValid());
+        $this->assertFalse($rule->validate('!2008:fa::0:1/0')->isValid());
+        $this->assertFalse($rule->validate('192.168.5.32/33')->isValid());
+        $this->assertFalse($rule->validate('2008:fz::0/129')->isValid());
+        $this->assertFalse($rule->validate('192.168.5.32/33')->isValid());
+        $this->assertFalse($rule->validate('192.168.5.32/af')->isValid());
+        $this->assertFalse($rule->validate('192.168.5.32/11/12')->isValid());
+        $this->assertTrue($rule->validate('10.0.0.1/24')->isValid());
+        $this->assertTrue($rule->validate('10.0.0.1/0')->isValid());
+        $this->assertTrue($rule->validate('2008:db0::1/64')->isValid());
+        $this->assertFalse($rule->validate('2008:db0::1')->isValid());
+        $this->assertFalse($rule->validate('10.0.0.1')->isValid());
 
-        $validator = $validator->allowIpv6()->requireSubnet();
-        $this->assertTrue($validator->validate('192.168.5.32/11')->isValid());
-        $this->assertTrue($validator->validate('192.168.5.32/32')->isValid());
-        $this->assertTrue($validator->validate('0.0.0.0/0')->isValid());
-        $this->assertTrue($validator->validate('2008:fa::0:1/64')->isValid());
-        $this->assertTrue($validator->validate('2008:fa::0:1/128')->isValid());
-        $this->assertTrue($validator->validate('2008:fa::0:1/0')->isValid());
-        $this->assertFalse($validator->validate('!2008:fa::0:1/0')->isValid());
-        $this->assertFalse($validator->validate('192.168.5.32/33')->isValid());
-        $this->assertFalse($validator->validate('2008:fz::0/129')->isValid());
-        $this->assertFalse($validator->validate('192.168.5.32/33')->isValid());
-        $this->assertFalse($validator->validate('192.168.5.32/af')->isValid());
-        $this->assertFalse($validator->validate('192.168.5.32/11/12')->isValid());
-
-        $validator = $validator->requireSubnet();
-        $this->assertTrue($validator->validate('10.0.0.1/24')->isValid());
-        $this->assertTrue($validator->validate('10.0.0.1/0')->isValid());
-        $this->assertTrue($validator->validate('2008:db0::1/64')->isValid());
-        $this->assertFalse($validator->validate('2008:db0::1')->isValid());
-        $this->assertFalse($validator->validate('10.0.0.1')->isValid());
-
-        $validator = $validator->allowNegation();
-        $this->assertTrue($validator->validate('!192.168.5.32/32')->isValid());
-        $this->assertTrue($validator->validate('!2008:fa::0:1/64')->isValid());
-        $this->assertFalse($validator->validate('!!192.168.5.32/32')->isValid());
-        $this->assertFalse($validator->validate('!!2008:fa::0:1/64')->isValid());
+        $rule = new Ip(requireSubnet: true, allowNegation: true);
+        $this->assertTrue($rule->validate('!192.168.5.32/32')->isValid());
+        $this->assertTrue($rule->validate('!2008:fa::0:1/64')->isValid());
+        $this->assertFalse($rule->validate('!!192.168.5.32/32')->isValid());
+        $this->assertFalse($rule->validate('!!2008:fa::0:1/64')->isValid());
     }
 
     public function testValidateRangeIPv4(): void
     {
-        $validator = Ip::rule()->ranges(['10.0.1.0/24']);
-        $this->assertTrue($validator->validate('10.0.1.2')->isValid());
-        $this->assertFalse($validator->validate('192.5.1.1')->isValid());
+        $rule = new Ip(ranges: ['10.0.1.0/24']);
+        $this->assertTrue($rule->validate('10.0.1.2')->isValid());
+        $this->assertFalse($rule->validate('192.5.1.1')->isValid());
 
-        $validator = $validator->ranges(['10.0.1.0/24']);
-        $this->assertTrue($validator->validate('10.0.1.2')->isValid());
-        $this->assertFalse($validator->validate('10.0.3.2')->isValid());
+        $rule = new Ip(ranges: ['10.0.1.0/24']);
+        $this->assertTrue($rule->validate('10.0.1.2')->isValid());
+        $this->assertFalse($rule->validate('10.0.3.2')->isValid());
 
-        $validator = $validator->ranges(['!10.0.1.0/24', '10.0.0.0/8', 'localhost']);
-        $this->assertFalse($validator->validate('10.0.1.2')->isValid());
-        $this->assertTrue($validator->validate('127.0.0.1')->isValid());
+        $rule = new Ip(ranges: ['!10.0.1.0/24', '10.0.0.0/8', 'localhost']);
+        $this->assertFalse($rule->validate('10.0.1.2')->isValid());
+        $this->assertTrue($rule->validate('127.0.0.1')->isValid());
 
-        $validator = $validator->allowSubnet()->ranges(['10.0.1.0/24', '!10.0.0.0/8', 'localhost']);
-        $this->assertTrue($validator->validate('10.0.1.2')->isValid());
-        $this->assertTrue($validator->validate('127.0.0.1')->isValid());
-        $this->assertTrue($validator->validate('10.0.1.28/28')->isValid());
-        $this->assertFalse($validator->validate('10.2.2.2')->isValid());
-        $this->assertFalse($validator->validate('10.0.1.1/22')->isValid());
+        $rule = new Ip(allowSubnet: true, ranges: ['10.0.1.0/24', '!10.0.0.0/8', 'localhost']);
+        $this->assertTrue($rule->validate('10.0.1.2')->isValid());
+        $this->assertTrue($rule->validate('127.0.0.1')->isValid());
+        $this->assertTrue($rule->validate('10.0.1.28/28')->isValid());
+        $this->assertFalse($rule->validate('10.2.2.2')->isValid());
+        $this->assertFalse($rule->validate('10.0.1.1/22')->isValid());
     }
 
     public function testValidateRangeIPv6(): void
     {
-        $validator = Ip::rule()->ranges(['2001:db0:1:1::/64']);
-        $this->assertTrue($validator->validate('2001:db0:1:1::6')->isValid());
-        $this->assertFalse($validator->validate('2001:db0:1:2::7')->isValid());
+        $rule = new Ip(ranges: ['2001:db0:1:1::/64']);
+        $this->assertTrue($rule->validate('2001:db0:1:1::6')->isValid());
+        $this->assertFalse($rule->validate('2001:db0:1:2::7')->isValid());
 
-        $validator = $validator->ranges(['2001:db0:1:2::/64']);
-        $this->assertTrue($validator->validate('2001:db0:1:2::7')->isValid());
+        $rule = new Ip(ranges: ['2001:db0:1:2::/64']);
+        $this->assertTrue($rule->validate('2001:db0:1:2::7')->isValid());
 
-        $validator = $validator->ranges(['!2001:db0::/32', '2001:db0:1:2::/64']);
-        $this->assertFalse($validator->validate('2001:db0:1:2::7')->isValid());
+        $rule = new Ip(ranges: ['!2001:db0::/32', '2001:db0:1:2::/64']);
+        $this->assertFalse($rule->validate('2001:db0:1:2::7')->isValid());
 
-        $validator = $validator->allowSubnet()->ranges(array_reverse($validator->getRanges()));
-        $this->assertTrue($validator->validate('2001:db0:1:2::7')->isValid());
+        $rule = new Ip(allowSubnet: true, ranges: array_reverse($rule->getRanges()));
+        $this->assertTrue($rule->validate('2001:db0:1:2::7')->isValid());
     }
 
     public function testValidateRangeIPvBoth(): void
     {
-        $validator = Ip::rule()->ranges(['10.0.1.0/24']);
-        $this->assertTrue($validator->validate('10.0.1.2')->isValid());
-        $this->assertFalse($validator->validate('192.5.1.1')->isValid());
-        $this->assertFalse($validator->validate('2001:db0:1:2::7')->isValid());
+        $rule = new Ip(ranges: ['10.0.1.0/24']);
+        $this->assertTrue($rule->validate('10.0.1.2')->isValid());
+        $this->assertFalse($rule->validate('192.5.1.1')->isValid());
+        $this->assertFalse($rule->validate('2001:db0:1:2::7')->isValid());
 
-        $validator = $validator->ranges(['10.0.1.0/24', '2001:db0:1:2::/64', '127.0.0.1']);
-        $this->assertTrue($validator->validate('2001:db0:1:2::7')->isValid());
-        $this->assertTrue($validator->validate('10.0.1.2')->isValid());
-        $this->assertFalse($validator->validate('10.0.3.2')->isValid());
+        $rule = new Ip(ranges: ['10.0.1.0/24', '2001:db0:1:2::/64', '127.0.0.1']);
+        $this->assertTrue($rule->validate('2001:db0:1:2::7')->isValid());
+        $this->assertTrue($rule->validate('10.0.1.2')->isValid());
+        $this->assertFalse($rule->validate('10.0.3.2')->isValid());
 
-        $validator = $validator->ranges(['!system', 'any']);
-        $this->assertFalse($validator->validate('127.0.0.1')->isValid());
-        $this->assertFalse($validator->validate('fe80::face')->isValid());
-        $this->assertTrue($validator->validate('8.8.8.8')->isValid());
+        $rule = new Ip(ranges: ['!system', 'any']);
+        $this->assertFalse($rule->validate('127.0.0.1')->isValid());
+        $this->assertFalse($rule->validate('fe80::face')->isValid());
+        $this->assertTrue($rule->validate('8.8.8.8')->isValid());
 
-        $validator = $validator->allowSubnet()->ranges(['10.0.1.0/24', '2001:db0:1:2::/64', 'localhost', '!any']);
-        $this->assertTrue($validator->validate('10.0.1.2')->isValid());
-        $this->assertTrue($validator->validate('2001:db0:1:2::7')->isValid());
-        $this->assertTrue($validator->validate('127.0.0.1')->isValid());
-        $this->assertTrue($validator->validate('10.0.1.28/28')->isValid());
-        $this->assertFalse($validator->validate('10.2.2.2')->isValid());
-        $this->assertFalse($validator->validate('10.0.1.1/22')->isValid());
+        $rule = new Ip(allowSubnet: true, ranges: ['10.0.1.0/24', '2001:db0:1:2::/64', 'localhost', '!any']);
+        $this->assertTrue($rule->validate('10.0.1.2')->isValid());
+        $this->assertTrue($rule->validate('2001:db0:1:2::7')->isValid());
+        $this->assertTrue($rule->validate('127.0.0.1')->isValid());
+        $this->assertTrue($rule->validate('10.0.1.28/28')->isValid());
+        $this->assertFalse($rule->validate('10.2.2.2')->isValid());
+        $this->assertFalse($rule->validate('10.0.1.1/22')->isValid());
     }
 
     public function testIpv4LeadingZero(): void
     {
-        $this->assertFalse(Ip::rule()->validate('01.01.01.01')->isValid());
-        $this->assertFalse(Ip::rule()->validate('010.010.010.010')->isValid());
-        $this->assertFalse(Ip::rule()->validate('001.001.001.001')->isValid());
+        $this->assertFalse((new Ip())->validate('01.01.01.01')->isValid());
+        $this->assertFalse((new Ip())->validate('010.010.010.010')->isValid());
+        $this->assertFalse((new Ip())->validate('001.001.001.001')->isValid());
     }
 
     public function testNetworkAlias(): void
     {
-        $validator = Ip::rule()->network('myNetworkEu', ['1.2.3.4/10', '5.6.7.8'])
-            ->ranges(['myNetworkEu']);
-        $this->assertTrue($validator->validate('1.2.3.4')->isValid());
-        $this->assertTrue($validator->validate('5.6.7.8')->isValid());
+        $rule = new Ip(networks: ['myNetworkEu' => ['1.2.3.4/10', '5.6.7.8']], ranges: ['myNetworkEu']);
+
+        $this->assertTrue($rule->validate('1.2.3.4')->isValid());
+        $this->assertTrue($rule->validate('5.6.7.8')->isValid());
     }
 
     public function testNetworkAliasException(): void
     {
-        $validator = Ip::rule()->network('myNetworkEu', ['1.2.3.4/10', '5.6.7.8'])
-            ->ranges(['myNetworkEu']);
+        $this->expectException(RuntimeException::class);
+        $this->expectExceptionMessage('Network alias "*" already set as default');
 
-        $this->expectException(\RuntimeException::class);
-        $this->expectExceptionMessage('Network alias "myNetworkEu" already set');
-        $validator->network('myNetworkEu', ['!1.2.3.4/10', '!5.6.7.8']);
+        new Ip(networks: ['*' => ['wrong']], ranges: ['*']);
     }
 
     public function testName(): void
     {
-        $this->assertEquals('ip', Ip::rule()->getName());
+        $this->assertEquals('ip', (new Ip())->getName());
     }
 
     public function optionsProvider(): array
     {
         return [
             [
-                Ip::rule(),
+                new Ip(),
                 [
-                    'message' => 'Must be a valid IP address.',
                     'allowIpv4' => true,
-                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
                     'allowIpv6' => true,
-                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
                     'allowSubnet' => false,
-                    'hasSubnetMessage' => 'Must not be a subnet.',
                     'requireSubnet' => false,
-                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
                     'allowNegation' => false,
+                    'message' => 'Must be a valid IP address.',
+                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
+                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
                     'wrongCidrMessage' => 'Contains wrong subnet mask.',
-                    'ranges' => [],
+                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
+                    'hasSubnetMessage' => 'Must not be a subnet.',
                     'notInRangeMessage' => 'Is not in the allowed range.',
+                    'ranges' => [],
                     'skipOnEmpty' => false,
                     'skipOnError' => false,
                 ],
             ],
             [
-                Ip::rule()->allowIpv4(),
+                new Ip(allowIpv4: false),
                 [
-                    'message' => 'Must be a valid IP address.',
-                    'allowIpv4' => true,
-                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
-                    'allowIpv6' => true,
-                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
-                    'allowSubnet' => false,
-                    'hasSubnetMessage' => 'Must not be a subnet.',
-                    'requireSubnet' => false,
-                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
-                    'allowNegation' => false,
-                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
-                    'ranges' => [],
-                    'notInRangeMessage' => 'Is not in the allowed range.',
-                    'skipOnEmpty' => false,
-                    'skipOnError' => false,
-                ],
-            ],
-            [
-                Ip::rule()->disallowIpv4(),
-                [
-                    'message' => 'Must be a valid IP address.',
                     'allowIpv4' => false,
-                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
                     'allowIpv6' => true,
-                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
                     'allowSubnet' => false,
-                    'hasSubnetMessage' => 'Must not be a subnet.',
-                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
                     'requireSubnet' => false,
-                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
                     'allowNegation' => false,
-                    'ranges' => [],
+                    'message' => 'Must be a valid IP address.',
+                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
+                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
+                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
+                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
+                    'hasSubnetMessage' => 'Must not be a subnet.',
                     'notInRangeMessage' => 'Is not in the allowed range.',
+                    'ranges' => [],
                     'skipOnEmpty' => false,
                     'skipOnError' => false,
                 ],
             ],
             [
-                Ip::rule()->disallowIpv6(),
+                new Ip(allowIpv6: false),
                 [
-                    'message' => 'Must be a valid IP address.',
                     'allowIpv4' => true,
-                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
                     'allowIpv6' => false,
-                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
                     'allowSubnet' => false,
-                    'hasSubnetMessage' => 'Must not be a subnet.',
-                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
                     'requireSubnet' => false,
-                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
                     'allowNegation' => false,
-                    'ranges' => [],
+                    'message' => 'Must be a valid IP address.',
+                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
+                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
+                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
+                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
+                    'hasSubnetMessage' => 'Must not be a subnet.',
                     'notInRangeMessage' => 'Is not in the allowed range.',
+                    'ranges' => [],
                     'skipOnEmpty' => false,
                     'skipOnError' => false,
                 ],
             ],
             [
-                Ip::rule()->allowSubnet(),
+                new Ip(allowSubnet: true),
                 [
-                    'message' => 'Must be a valid IP address.',
                     'allowIpv4' => true,
-                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
                     'allowIpv6' => true,
-                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
                     'allowSubnet' => true,
-                    'hasSubnetMessage' => 'Must not be a subnet.',
-                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
                     'requireSubnet' => false,
-                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
                     'allowNegation' => false,
-                    'ranges' => [],
+                    'message' => 'Must be a valid IP address.',
+                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
+                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
+                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
+                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
+                    'hasSubnetMessage' => 'Must not be a subnet.',
                     'notInRangeMessage' => 'Is not in the allowed range.',
+                    'ranges' => [],
                     'skipOnEmpty' => false,
                     'skipOnError' => false,
                 ],
             ],
             [
-                Ip::rule()->requireSubnet(),
+                new Ip(requireSubnet: true),
                 [
-                    'message' => 'Must be a valid IP address.',
                     'allowIpv4' => true,
-                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
                     'allowIpv6' => true,
-                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
                     'allowSubnet' => true,
-                    'hasSubnetMessage' => 'Must not be a subnet.',
-                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
                     'requireSubnet' => true,
-                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
                     'allowNegation' => false,
-                    'ranges' => [],
+                    'message' => 'Must be a valid IP address.',
+                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
+                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
+                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
+                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
+                    'hasSubnetMessage' => 'Must not be a subnet.',
                     'notInRangeMessage' => 'Is not in the allowed range.',
+                    'ranges' => [],
                     'skipOnEmpty' => false,
                     'skipOnError' => false,
                 ],
             ],
             [
-                Ip::rule()->requireSubnet()->disallowSubnet(),
+                new Ip(allowNegation: true),
                 [
-                    'message' => 'Must be a valid IP address.',
                     'allowIpv4' => true,
-                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
                     'allowIpv6' => true,
-                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
                     'allowSubnet' => false,
-                    'hasSubnetMessage' => 'Must not be a subnet.',
-                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
                     'requireSubnet' => false,
-                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
-                    'allowNegation' => false,
-                    'ranges' => [],
-                    'notInRangeMessage' => 'Is not in the allowed range.',
-                    'skipOnEmpty' => false,
-                    'skipOnError' => false,
-                ],
-            ],
-            [Ip::rule()->allowNegation(),
-                [
-                    'message' => 'Must be a valid IP address.',
-                    'allowIpv4' => true,
-                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
-                    'allowIpv6' => true,
-                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
-                    'allowSubnet' => false,
-                    'hasSubnetMessage' => 'Must not be a subnet.',
-                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
-                    'requireSubnet' => false,
-                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
                     'allowNegation' => true,
-                    'ranges' => [],
+                    'message' => 'Must be a valid IP address.',
+                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
+                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
+                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
+                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
+                    'hasSubnetMessage' => 'Must not be a subnet.',
                     'notInRangeMessage' => 'Is not in the allowed range.',
+                    'ranges' => [],
                     'skipOnEmpty' => false,
                     'skipOnError' => false,
                 ],
             ],
             [
-                Ip::rule()->allowNegation()->disallowNegation(),
+                new Ip(ranges: ['private']),
                 [
-                    'message' => 'Must be a valid IP address.',
                     'allowIpv4' => true,
-                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
                     'allowIpv6' => true,
-                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
                     'allowSubnet' => false,
-                    'hasSubnetMessage' => 'Must not be a subnet.',
-                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
                     'requireSubnet' => false,
-                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
                     'allowNegation' => false,
-                    'ranges' => [],
+                    'message' => 'Must be a valid IP address.',
+                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
+                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
+                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
+                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
+                    'hasSubnetMessage' => 'Must not be a subnet.',
                     'notInRangeMessage' => 'Is not in the allowed range.',
-                    'skipOnEmpty' => false,
-                    'skipOnError' => false,
-                ],
-            ],
-            [
-                Ip::rule()->ranges(['private']),
-                [
-                    'message' => 'Must be a valid IP address.',
-                    'allowIpv4' => true,
-                    'ipv4NotAllowedMessage' => 'Must not be an IPv4 address.',
-                    'allowIpv6' => true,
-                    'ipv6NotAllowedMessage' => 'Must not be an IPv6 address.',
-                    'allowSubnet' => false,
-                    'hasSubnetMessage' => 'Must not be a subnet.',
-                    'wrongCidrMessage' => 'Contains wrong subnet mask.',
-                    'requireSubnet' => false,
-                    'noSubnetMessage' => 'Must be an IP address with specified subnet.',
-                    'allowNegation' => false,
                     'ranges' => ['10.0.0.0/8', '172.16.0.0/12', '192.168.0.0/16', 'fd00::/8'],
-                    'notInRangeMessage' => 'Is not in the allowed range.',
                     'skipOnEmpty' => false,
                     'skipOnError' => false,
                 ],
